trader_orchestrator_phase_next/requirements.txt
flask
python-dotenv
requests
numpy
pandas
scikit-learn==1.4.2
joblib
